Add: JavaDocs for package commands
