+++ src/main/java/crayon/commands/AddToDoCommand.java
@@ -1,4 +1,4 @@
-package crayon.command;
+package crayon.commands;
 
 import crayon.Storage;
 import crayon.TaskList;
@@ -6,15 +6,33 @@ import crayon.Ui;
 import crayon.enums.Action;
 import crayon.enums.TaskType;
 
+/**
+ * Represents a command to add a todo task.
+ */
 public class AddToDoCommand extends Command{
 
+    /**
+     * The content of the todo task.
+     */
     private final String content;
 
+    /**
+     * Constructs an AddToDoCommand.
+     *
+     * @param content The content of the todo task.
+     */
     public AddToDoCommand(String content) {
         super(Action.TODO);
         this.content = content;
     }
 
+    /**
+     * Executes the command to add a todo task.
+     *
+     * @param storage The storage to save the task.
+     * @param ui The user interface to show messages.
+     * @param taskList The task list to add the task.
+     */
     @Override
     public void execute(Storage storage, Ui ui, TaskList taskList) {
         String todoMessage = taskList.createTask(TaskType.TODO, content);
