+++ src/main/java/crayon/commands/DeleteCommand.java
@@ -1,4 +1,4 @@
-package crayon.command;
+package crayon.commands;
 
 import crayon.Storage;
 import crayon.TaskList;
@@ -6,10 +6,18 @@ import crayon.Ui;
 import crayon.enums.Action;
 import crayon.exceptions.CrayonIllegalArgumentException;
 
+/**
+ * Represents a command to delete a task.
+ */
 public class DeleteCommand extends Command {
 
     private final String content;
 
+    /**
+     * Constructs a DeleteCommand.
+     *
+     * @param content The content of the task to be deleted.
+     */
     public DeleteCommand(String content) {
         super(Action.DELETE);
         this.content = content;
